adjust some new things
