+++ src/main/resources/templates/courseDetail.html
@@ -151,7 +151,7 @@
         </div>
     </div>
     <div id="reviewModal" class="modal">
-        <div class="modal-content">
+        <div class="review-modal-content">
             <span class="close-btn" onclick="closeReviewPopup()">&times;</span>
             <h2>Submit Your Review</h2>
