+++ src/main/resources/templates/fragments/header_nav.html
@@ -19,9 +19,6 @@
             <input name="title" type="text" placeholder="Search my courses" th:value="${param.title}">
         </form>
         <div class="user-actions">
-            <a class="action" href="">
-                <i class="fa fa-gift"></i>
-            </a>
             <a class="action" href="/show-wish-list">
                 <i class="fa-solid fa-heart"></i>
             </a>
