--- src/main/resources/templates/giftCart.html
@@ -1,114 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.w3.org/1999/xhtml">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css">
-    <title>Gift Cart</title>
-    <link rel="stylesheet" th:href="@{/css/style.css}">
-</head>
-<body>
-
-    <h1 class="cart-header title">Gift Cart</h1>
-    <p class="cart-header course-num">2 courses in cart</p>
-    <div class="receiver-list-container">
-        <p class="cart-header course-num">Receiver:</p>
-
-        <!-- For loop to get all users or friends(if have) -->
-        <select class="receiver-list">
-            <option value="abc@gmail.com">abc@gmail.com</option>
-            <option value="abc@gmail.com">abc@gmail.com</option>
-            <option value="abc@gmail.com">abc@gmail.com</option>
-        </select>
-    </div>
-    <!-- CART TABLE -->
-    <div class="cart-section">
-        <div class="cart-container">
-
-            <!-- For loop to get the cart items -->
-            <div class="cart-item">
-                <a class="cart-img-container" href="">
-                    <img src="https://cdn.bap-software.net/2024/08/26213247/spring.jpg" alt="">
-                </a>
-                <div class="cart-text-container">
-                    <a class="related-course-name">Linear algebra and its applications</a>
-                    <p class="author">Nguyen Tung Anh</p>
-                    <p class="star"><i class="fa-solid fa-star"></i>4,8</p>
-                    <div class="remove-container">
-                        <button class="remove-btn">Remove</button>
-                        <p class="related-course-price">đ 2500000</p>
-                    </div>
-                </div>
-            </div>
-            <div class="cart-item">
-                <a class="cart-img-container" href="">
-                    <img src="https://cdn.bap-software.net/2024/08/26213247/spring.jpg" alt="">
-                </a>
-                <div class="cart-text-container">
-                    <a class="related-course-name">Linear algebra and its applications</a>
-                    <p class="author">Nguyen Tung Anh</p>
-                    <p class="star"><i class="fa-solid fa-star"></i>4,8</p>
-                    <div class="remove-container">
-                        <button class="remove-btn">Remove</button>
-                        <p class="related-course-price">đ 2500000</p>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-
-        <!-- Puchasing section -->
-        <div class="cart-purchasing-section">
-            <h2>Total:</h2>
-            <p class="price">đ 2.500.000</p>
-            <button class="btn purchasing-button">Send to friend</button>
-        </div>
-    </div>
-
-
-    <!-- RELATED COURSES -->
-    <div class="related-container cart-related-container">
-        <div class="related-header">
-            Related courses
-        </div>
-        <div class="course-container">
-            <div class="course">
-                <div class="img-container">
-                    <img src="https://cdn.bap-software.net/2024/08/26213247/spring.jpg" alt="">
-                </div>
-                <a class="related-course-name">Linear algebra and its applications</a>
-                <p class="author">Nguyen Tung Anh</p>
-                <p class="star"><i class="fa-solid fa-star"></i>4,8</p>
-                <p class="related-course-price">đ 2500000</p>
-            </div>
-            <div class="course">
-                <div class="img-container">
-                    <img src="https://cdn.bap-software.net/2024/08/26213247/spring.jpg" alt="">
-                </div>
-                <a class="related-course-name">Linear algebra and its applications</a>
-                <p class="author">Nguyen Tung Anh</p>
-                <p class="star"><i class="fa-solid fa-star"></i>4,8</p>
-                <p class="related-course-price">đ 2500000</p>
-            </div>
-            <div class="course">
-                <div class="img-container">
-                    <img src="https://cdn.bap-software.net/2024/08/26213247/spring.jpg" alt="">
-                </div>
-                <a class="related-course-name">Linear algebra and its applications</a>
-                <p class="author">Nguyen Tung Anh</p>
-                <p class="star"><i class="fa-solid fa-star"></i>4,8</p>
-                <p class="related-course-price">đ 2500000</p>
-            </div>
-            <div class="course">
-                <div class="img-container">
-                    <img src="https://cdn.bap-software.net/2024/08/26213247/spring.jpg" alt="">
-                </div>
-                <a class="related-course-name">Linear algebra and its applications</a>
-                <p class="author">Nguyen Tung Anh</p>
-                <p class="star"><i class="fa-solid fa-star"></i>4,8</p>
-                <p class="related-course-price">đ 2500000</p>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
