+++ src/main/resources/templates/homePage1.html
@@ -146,9 +146,9 @@
             <div class="about-info">
                 <img src="https://edunow.vn/wp-content/uploads/2021/05/elearning-la-gi-06.jpg" alt="">
                 <div class="inner-section">
-                    <p style="font-size: 15px; align-items: center;" class="text">
-                        <strong style="font-size: 18px; margin-bottom: 10px;">Empowering Your Journey, Anytime,
-                            Anywhere</strong><br>
+                    <h3 style="margin-bottom: 15px;font-size: 20px;"><strong style="font-size: 18px; margin-bottom: 10px;">Empowering Your Journey, Anytime,
+                        Anywhere</strong></h3>
+                    <p style="font-size: 18px; align-items: center;" class="text">
                         At LearniVerse, we believe that education should be accessible, engaging, and tailored to your
                         needs. Whether you're looking to upskill, explore a new career, or follow your passion, our
                         platform connects you with expert instructors and top-quality courses designed for real-world
